requirements.txt
gunicorn==21.2.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
paramiko==3.3.1
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
scp==0.14.5

